fix clear removed files

+++ Deploy/deploy.cpp
@@ -1064,6 +1064,10 @@ void Deploy::clear(bool force) {
 
         auto index = it - 1;
 
+        if (!index.value().exists()) {
+            continue;
+        }
+
         if (index.value().isFile()) {
             if (removeFile(index.value())) {
                 qInfo() << "Remove " << index.value().absoluteFilePath() << " becouse it is deployed file";
